# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Order Microservice API",
    "description": "API contract for the Order Microservice. Handles order creation, retrieval, and status management for a mobile accessory e-commerce platform. All endpoints require JWT Bearer authentication.",
    "version": "1.0.0",
    "contact": {
      "name": "Ishan Dilmith",
      "url": "https://github.com/IshanDilmith/order-microservice"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8003",
      "description": "Local Development Server"
    },
    {
      "url": "https://api-gateway.wonderfulpond-b48eec63.southeastasia.azurecontainerapps.io",
      "description": "Production Gateway (Azure)"
    }
  ],
  "tags": [
    {
      "name": "Orders",
      "description": "Order management operations"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Orders"],
        "summary": "Service health check",
        "description": "Returns a simple message confirming that the order service is running.",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order service is running"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/create": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create a new order",
        "description": "Creates a new order for the authenticated user. Fetches cart data via /cart/{userId} and resolves the user email via /auth/users/{userId} through the API gateway, decrements inventory for each item, saves the order with an auto-generated custom order ID (e.g. ORD#0001), and sends an email notification.",
        "operationId": "createOrder",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              },
              "example": {
                "payMethod": "COD",
                "deliveryAddress": "123 Main Street, Colombo",
                "province": "Western",
                "district": "Colombo",
                "totalDiscount": 150,
                "deliveryFee": 350
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderResponse"
                },
                "example": {
                  "success": true,
                  "order": {
                    "_id": "665a1b2c3d4e5f6a7b8c9d0e",
                    "orderId": "ORD#0001",
                    "userId": "user123",
                    "userEmail": "user@example.com",
                    "items": [
                      {
                        "productId": "prod001",
                        "productName": "Wireless Charger",
                        "productImage": {
                          "url": "https://cdn.example.com/products/prod001/image-1.jpg",
                          "publicId": "products/prod001/image-1"
                        },
                        "quantity": 2,
                        "price": 1500
                      }
                    ],
                    "deliveryAddress": "123 Main Street, Colombo",
                    "province": "Western",
                    "district": "Colombo",
                    "totalDiscount": 150,
                    "deliveryFee": 350,
                    "total": 3200,
                    "status": "Pending",
                    "payMethod": "COD",
                    "createdAt": "2026-03-21T06:30:00.000Z"
                  },
                  "notification": {
                    "success": true,
                    "message": "Notification sent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request — Cart is empty or user email not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "emptyCart": {
                    "summary": "Cart is empty",
                    "value": { "error": "Cart is empty" }
                  },
                  "noEmail": {
                    "summary": "User email not found",
                    "value": { "error": "User email not found" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error — Failed to create order or update inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createFailed": {
                    "summary": "Order creation failed",
                    "value": { "error": "Failed to create order" }
                  },
                  "inventoryFailed": {
                    "summary": "Inventory update failed",
                    "value": { "error": "Failed to update inventory for one or more items." }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Update order status",
        "description": "Updates the status of an existing order identified by its MongoDB `_id`. After updating, the service resolves the user's email via /auth/users/{userId} and sends a notification email through the gateway.",
        "operationId": "updateOrderStatus",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the order",
            "schema": {
              "type": "string",
              "example": "665a1b2c3d4e5f6a7b8c9d0e"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              },
              "example": {
                "status": "Confirmed"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateStatusResponse"
                },
                "example": {
                  "success": true,
                  "order": {
                    "_id": "665a1b2c3d4e5f6a7b8c9d0e",
                    "orderId": "ORD#0001",
                    "userId": "user123",
                    "items": [
                      {
                        "productId": "prod001",
                        "quantity": 2,
                        "price": 1500
                      }
                    ],
                    "deliveryAddress": "123 Main Street, Colombo",
                    "province": "Western",
                    "district": "Colombo",
                    "totalDiscount": 150,
                    "deliveryFee": 350,
                    "total": 3200,
                    "status": "Confirmed",
                    "payMethod": "COD",
                    "createdAt": "2026-03-21T06:30:00.000Z"
                  },
                  "notification": {
                    "success": true,
                    "message": "Notification sent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request — User email not found for notification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": { "error": "User email not found" }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found — No order with the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": { "error": "Order not found" }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": { "error": "Failed to update status" }
              }
            }
          }
        }
      }
    },
    "/orders/my-orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get current user's orders",
        "description": "Retrieves all orders belonging to the authenticated user (identified via the JWT token).",
        "operationId": "getUserOrders",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                },
                "example": {
                  "success": true,
                  "orders": [
                    {
                      "_id": "665a1b2c3d4e5f6a7b8c9d0e",
                      "orderId": "ORD#0001",
                      "userId": "user123",
                      "items": [
                        {
                          "productId": "prod001",
                          "quantity": 2,
                          "price": 1500
                        }
                      ],
                      "deliveryAddress": "123 Main Street, Colombo",
                      "province": "Western",
                      "district": "Colombo",
                      "totalDiscount": 150,
                      "deliveryFee": 350,
                      "total": 3200,
                      "status": "Pending",
                      "payMethod": "COD",
                      "createdAt": "2026-03-21T06:30:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": { "error": "Failed to fetch orders" }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID",
        "description": "Retrieves a single order by its MongoDB `_id`.",
        "operationId": "getOrderById",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the order",
            "schema": {
              "type": "string",
              "example": "665a1b2c3d4e5f6a7b8c9d0e"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleOrderResponse"
                },
                "example": {
                  "success": true,
                  "order": {
                    "_id": "665a1b2c3d4e5f6a7b8c9d0e",
                    "orderId": "ORD#0001",
                    "userId": "user123",
                    "items": [
                      {
                        "productId": "prod001",
                        "quantity": 2,
                        "price": 1500
                      }
                    ],
                    "deliveryAddress": "123 Main Street, Colombo",
                    "province": "Western",
                    "district": "Colombo",
                    "totalDiscount": 150,
                    "deliveryFee": 350,
                    "total": 3200,
                    "status": "Pending",
                    "payMethod": "COD",
                    "createdAt": "2026-03-21T06:30:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found — No order with the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": { "error": "Order not found" }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": { "error": "Failed to fetch order" }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get all orders",
        "description": "Retrieves all orders in the system.",
        "operationId": "getAllOrders",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "All orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                },
                "example": {
                  "success": true,
                  "orders": [
                    {
                      "_id": "665a1b2c3d4e5f6a7b8c9d0e",
                      "orderId": "ORD#0001",
                      "userId": "user123",
                      "items": [
                        {
                          "productId": "prod001",
                          "quantity": 2,
                          "price": 1500
                        }
                      ],
                      "deliveryAddress": "123 Main Street, Colombo",
                      "province": "Western",
                      "district": "Colombo",
                      "totalDiscount": 150,
                      "deliveryFee": 350,
                      "total": 3200,
                      "status": "Pending",
                      "payMethod": "COD",
                      "createdAt": "2026-03-21T06:30:00.000Z"
                    },
                    {
                      "_id": "665a1b2c3d4e5f6a7b8c9d1f",
                      "orderId": "ORD#0002",
                      "userId": "user456",
                      "items": [
                        {
                          "productId": "prod002",
                          "quantity": 1,
                          "price": 2500
                        }
                      ],
                      "deliveryAddress": "456 Oak Avenue, Kandy",
                      "province": "Central",
                      "district": "Kandy",
                      "totalDiscount": 0,
                      "deliveryFee": 500,
                      "total": 3000,
                      "status": "Confirmed",
                      "payMethod": "Card",
                      "createdAt": "2026-03-21T07:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": { "error": "Failed to fetch orders" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from the authentication service. Format: `Bearer <token>`"
      }
    },
    "schemas": {
      "OrderItem": {
        "type": "object",
        "required": ["productId", "productName", "quantity", "price"],
        "properties": {
          "productId": {
            "type": "string",
            "description": "Unique identifier of the product",
            "example": "prod001"
          },
          "productName": {
            "type": "string",
            "description": "Name of the product",
            "example": "Wireless Charger"
          },
          "productImage": {
            "type": "object",
            "nullable": true,
            "description": "First image of the product",
            "properties": {
              "url": {
                "type": "string",
                "example": "https://cdn.example.com/products/prod001/image-1.jpg"
              },
              "publicId": {
                "type": "string",
                "example": "products/prod001/image-1"
              }
            }
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Quantity ordered",
            "example": 2
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "description": "Unit price of the product",
            "example": 1500
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB-generated unique identifier",
            "example": "665a1b2c3d4e5f6a7b8c9d0e"
          },
          "orderId": {
            "type": "string",
            "description": "Custom human-readable order identifier (e.g. ORD#0001)",
            "example": "ORD#0001"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who placed the order",
            "example": "user123"
          },
          "userEmail": {
            "type": "string",
            "description": "Email address of the user who placed the order",
            "example": "user@example.com"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "description": "List of items in the order"
          },
          "deliveryAddress": {
            "type": "string",
            "description": "Full delivery address",
            "example": "123 Main Street, Colombo"
          },
          "province": {
            "type": "string",
            "description": "Province for delivery",
            "example": "Western"
          },
          "district": {
            "type": "string",
            "description": "District for delivery",
            "example": "Colombo"
          },
          "totalDiscount": {
            "type": "number",
            "format": "double",
            "default": 0,
            "description": "Total discount applied to the order",
            "example": 150
          },
          "deliveryFee": {
            "type": "number",
            "format": "double",
            "default": 0,
            "description": "Delivery fee charged for the order",
            "example": 350
          },
          "total": {
            "type": "number",
            "format": "double",
            "description": "Final total = sum(item.price × item.quantity) − totalDiscount + deliveryFee",
            "example": 3200
          },
          "status": {
            "type": "string",
            "enum": ["Pending", "Confirmed", "Notified", "Failed"],
            "description": "Current status of the order",
            "example": "Pending"
          },
          "payMethod": {
            "type": "string",
            "enum": ["COD", "Card"],
            "description": "Payment method selected by the user",
            "example": "COD"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was created",
            "example": "2026-03-21T06:30:00.000Z"
          },
          "notifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Timestamp when the user was notified about the order",
            "example": null
          }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "required": ["payMethod", "deliveryAddress", "province", "district"],
        "properties": {
          "payMethod": {
            "type": "string",
            "enum": ["COD", "Card"],
            "description": "Payment method",
            "example": "COD"
          },
          "deliveryAddress": {
            "type": "string",
            "description": "Full delivery address",
            "example": "123 Main Street, Colombo"
          },
          "province": {
            "type": "string",
            "description": "Province for delivery",
            "example": "Western"
          },
          "district": {
            "type": "string",
            "description": "District for delivery",
            "example": "Colombo"
          },
          "totalDiscount": {
            "type": "number",
            "format": "double",
            "default": 0,
            "description": "Optional discount amount to apply",
            "example": 150
          },
          "deliveryFee": {
            "type": "number",
            "format": "double",
            "default": 0,
            "description": "Optional delivery fee to add",
            "example": 350
          }
        }
      },
      "UpdateStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["Pending", "Confirmed", "Notified", "Failed"],
            "description": "New status for the order",
            "example": "Confirmed"
          }
        }
      },
      "CreateOrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "notification": {
            "type": "object",
            "nullable": true,
            "description": "Response from the notification service",
            "properties": {
              "success": {
                "type": "boolean"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "UpdateStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "notification": {
            "type": "object",
            "nullable": true,
            "description": "Response from the notification service",
            "properties": {
              "success": {
                "type": "boolean"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "SingleOrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
      "OrderListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Authorization error message",
            "example": "Not authorized, token missing"
          },
          "error": {
            "type": "string",
            "description": "Detailed error (present for invalid tokens)",
            "example": "jwt malformed"
          }
        }
      }
    }
  }
}